requirements.txt
Django
transformers
torch
numpy
onnxruntime
python-dotenv
langchain_community
langchain_huggingface
pinecone
langchain-pinecone
langchain_groq
better_profanity
python-logging-loki
django-cors-headers
djangorestframework
gunicorn
sentence_transformers
upstash_redis
